requirements.txt
flask==2.2.2
PyYAML==6.0
markdown2[all]~=2.4.0
Flask-Minify==0.41
Pygments==2.13.0